Make sidebar width responsive
